Add base AIController / Move camera follow logice to CamFollow script
The AIController now uses the same inputs as the player to maneuver the aircraft towards a desired location. I had to modify the control logic for yaw to get this to work. Removed the camera logic from the PlayerFinal script and set up the code in CamFollow to mirror that logic. Added a more robust gitignore file.